+++ atm/results.py
@@ -1,3 +1,10 @@
+"""
+Results
+-------
+results.py
+
+Tools to generate summary of results
+"""
 import numpy as np
 import os, sys
 import datetime, time
@@ -15,94 +22,115 @@ def as_string(atm_obj, start_time, end_time):
     long_divider = 69
     short_divider = 35
     string = ''
-    string +=  ' '+ nl   
-    string +=  ' '+ nl
-    string +=  '=' * long_divider + nl
-    string +=  '       Simulation Results         '+ nl
-    string +=  '=' * long_divider+ nl
-    string +=  'ATM version: '+str(__version__)+ nl
-    string +=  'Code URL: '+str(__codeurl__)+ nl
-    string +=  'Simulation name: '+ atm_obj.control['Simulation_name']+ nl
-    string +=  'Start Date / Time : '+str( start_time)+ nl
-    string +=  'End Date / Time : '+str(end_time)+ nl
+    string += ' '+ nl   
+    string += ' '+ nl
+    string += '=' * long_divider + nl
+    string += '       Simulation Results         '+ nl
+    string += '=' * long_divider+ nl
+    string += 'ATM version: '+str(__version__)+ nl
+    string += 'Code URL: '+str(__codeurl__)+ nl
+    string += 'Simulation name: '+ atm_obj.control['Simulation_name']+ nl
+    string += 'Start Date / Time : '+str( start_time)+ nl
+    string += 'End Date / Time : '+str(end_time)+ nl
     duration = (end_time - start_time)
-    string +=  'Total Simulation Time (minutes): '+ str(duration) + nl
-    string +=  'Total Simulation Time (seconds): '+ str(duration.total_seconds()) + nl
-    string +=  'Number of time steps in the simulation: '+ str(atm_obj.stop)+ nl
-    string +=  'Time(Seconds)/Timestemp: ' + str(duration.total_seconds()/atm_obj.stop) + nl
+    string += 'Total Simulation Time (minutes): '+ str(duration) + nl
+    string += \
+        'Total Simulation Time (seconds): '+ str(duration.total_seconds()) + nl
+    string += \
+        'Number of time steps in the simulation: '+ str(atm_obj.stop)+ nl
+    string += \
+        'Time(Seconds)/Timestemp: ' + \
+        str(duration.total_seconds()/atm_obj.stop) + nl
 
     # return string
     if atm_obj.control['Archive_simulation']:
-        string +=  'Archive Status: Active'+ nl
-        string +=  'Archive Name : '+str(atm_obj.control['Simulation_name'])+ nl
+        string += 'Archive Status: Active'+ nl
+        string += 'Archive Name : '+str(atm_obj.control['Simulation_name'])+ nl
     else:
-        string +=  'Archive Status: Inactive'+ nl
-    string +=  ' '+ nl
-    string +=  'Number of time steps in the simulation: '+ str(atm_obj.stop)+ nl
-    string +=  ' '+ nl
-    string +=  '-' * short_divider+ nl
-    string +=  ' Initial Cohort Information'+ nl
-    string +=  '-' * short_divider+ nl
-    string +=  'Outputs of Initial Cohort Distribution:'+ nl
+        string += 'Archive Status: Inactive'+ nl
+    string += ' '+ nl
+    string += 'Number of time steps in the simulation: '+ str(atm_obj.stop)+ nl
+    string += ' '+ nl
+    string += '-' * short_divider+ nl
+    string += ' Initial Cohort Information'+ nl
+    string += '-' * short_divider+ nl
+    string += 'Outputs of Initial Cohort Distribution:'+ nl
     #~ print atm_obj.control['Initialize_Control']
-    if atm_obj.control['Initialize_Control']['Initial_Cohort_Distribution_Figure'] == False:
-        string +=  '  No outputs generated.'+ nl
+    icdf = atm_obj.control[
+        'Initialize_Control'
+    ]['Initial_Cohort_Distribution_Figure']
+    if icdf == False:
+        string += '  No outputs generated.'+ nl
     else:
         for c in atm_obj.control['Initialize_Control']:
-            if c.find('Figure') != -1 and atm_obj.control['Initialize_Control'][c]:
+            if c.find('Figure') != -1 and \
+                    atm_obj.control['Initialize_Control'][c]:
                 string += '  ' + c + nl
-    string +=  ' '+ nl
+    string += ' '+ nl
 
-    string +=  'Outputs of Normalized Cohort Distribution:'+ nl
-    if atm_obj.control['Initialize_Control']['Normalized_Cohort_Distribution_Figure'] == False:
-        string +=  '  No outputs generated.'+ nl
+    string += 'Outputs of Normalized Cohort Distribution:'+ nl
+    ncdf = atm_obj.control[
+        'Initialize_Control'
+    ]['Normalized_Cohort_Distribution_Figure']
+    if ncdf == False:
+        string += '  No outputs generated.'+ nl
     else:
         for c in atm_obj.control['Initialize_Control']:
-            if c.find('Normal') != -1 and atm_obj.control['Initialize_Control'][c]:
+            if c.find('Normal') != -1 and \
+                    atm_obj.control['Initialize_Control'][c]:
                 string += '  ' + c + nl
-    string +=  ' '+ nl
+    string += ' '+ nl
 
-    string +=  'Outputs of Cohort Ages:'+ nl
-    if atm_obj.control['Initialize_Control']['Initial_Cohort_Age_Figure'] == False:
-        string +=  '  No outputs generated.'+ nl
+    string += 'Outputs of Cohort Ages:'+ nl
+    icaf = atm_obj.control['Initialize_Control']['Initial_Cohort_Age_Figure'] 
+    if icaf == False:
+        string += '  No outputs generated.'+ nl
     else:
         for c in atm_obj.control['Initialize_Control']:
             if c.find('Age') != -1 and atm_obj.control['Initialize_Control'][c]:
                 string += '  ' + c + nl
 
-    string +=  ' '+ nl
+    string += ' '+ nl
     ##########################################################################
-    string +=  '-' * short_divider+ nl
-    string +=  '   Meteorologic Data Information   '+ nl
-    string +=  '-' * short_divider+ nl
+    string += '-' * short_divider+ nl
+    string += '   Meteorologic Data Information   '+ nl
+    string += '-' * short_divider+ nl
     # if atm_obj.control['Met_Control']['met_distribution'].lower() == 'point':
-    #     string +=  'Point meteorologic data is used.'+ nl
+    #     string += 'Point meteorologic data is used.'+ nl
     # else:
-    #     string +=  'Meteorologic data is distributed.'+ nl
-    # string +=  'Meteorologic Data File: '+ str(atm_obj.control['Met_Control']['met_file_distributed'])+ nl
+    #     string += 'Meteorologic data is distributed.'+ nl
+    # string += 'Meteorologic Data File: '+ \
+    #   str(atm_obj.control['Met_Control']['met_file_distributed'])+ nl
     # if atm_obj.control['Met_Control']['degree_day_method'].lower() == 'read':
-    #     string +=  'Degree Days read from files: ' + atm_obj.control['Met_Control']['TDD_file'] +' and '+atm_obj.control['Met_Control']['FDD_file']+ nl
+    #     string += 'Degree Days read from files: ' + \
+    #       atm_obj.control['Met_Control']['TDD_file'] +\
+    #       ' and '+atm_obj.control['Met_Control']['FDD_file']+ nl
     # else:
-    #     string +=  'Degree Days calculated during simulation.'+ nl
-    # string +=  ' '+ nl
+    #     string += 'Degree Days calculated during simulation.'+ nl
+    # string += ' '+ nl
 
-    string +=  'Outputs:'+ nl
+    string += 'Outputs:'+ nl
     if atm_obj.control['Met_Control']['Degree_Day_Output']:
-        string +=  '  Degree-Days are output.'+ nl
+        string += '  Degree-Days are output.'+ nl
         
     # Note: Might want to add climatic event probability and block size here
     ############################################################################
-    string +=  '-' * short_divider+ nl
-    string +=  '   General Terrestrial Information   '+ nl
-    string +=  '-' * short_divider+ nl
-    # string +=  'Ground Ice Distribution: '+atm_obj.control['Terrestrial_Control']['Ice_Distribution']+ nl
-    string +=  'Drainage Efficiency Distribution: '+ atm_obj.control['Terrestrial_Control']['Drainage_Efficiency_Distribution']+ nl
-    string +=  'Initial Active Layer Depth Distribution: '+atm_obj.control['Terrestrial_Control']['ALD_Distribution']+ nl
-    string +=  ' '+ nl
+    string += '-' * short_divider+ nl
+    string += '   General Terrestrial Information   ' + nl
+    string += '-' * short_divider+ nl
+    # string += 'Ground Ice Distribution: ' + \
+    #   atm_obj.control['Terrestrial_Control']['Ice_Distribution']+ nl
+    string += 'Drainage Efficiency Distribution: ' + \
+        atm_obj.control[
+            'Terrestrial_Control'
+        ]['Drainage_Efficiency_Distribution'] + nl
+    string += 'Initial Active Layer Depth Distribution: ' + \
+        atm_obj.control['Terrestrial_Control']['ALD_Distribution'] + nl
+    string += ' '+ nl
 
 
     start_year = atm_obj.control['start_year']
-    end_year =  atm_obj.control['initialization year'] + atm_obj.stop -1 
+    end_year =  atm_obj.control['initialization_year'] + atm_obj.stop -1 
     
     cohort_list = atm_obj.grids.area.key_to_index.keys()
     cohort_list = [c for c in cohort_list if c.find('--') == -1]
@@ -111,7 +139,7 @@ def as_string(atm_obj, start_time, end_time):
     nl = '\n'
     for cohort in sorted(cohort_list):
         cohort_type = cohort.split('_')[0]
-        #~ string +=  not (cohort_type in all_ages)
+        #~ string += not (cohort_type in all_ages)
         #~ string = ''
         if not (cohort_type in all_ages):
             string += '='*long_divider + nl